requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
python-dotenv